feat(): move debounceByAnimationFrame to @next-shared/general
