+++ bricks/data-view/package.json
@@ -41,6 +41,7 @@
     "@next-core/react-runtime": "^1.0.65",
     "@next-core/types": "^1.6.0",
     "@next-core/utils": "^1.5.7",
+    "@next-shared/general": "^0.3.13",
     "@tweenjs/tween.js": "^19.0.0",
     "classnames": "^2.3.2",
     "d3-hierarchy": "^3.1.2",
